No owner for TCell/LCell::new(). Add: owner.cell(), qcellowner.id()
TCell and LCell don't strictly need an owner reference to work; it was
included only for symmetry with QCell, and as an extra check.  However
requiring the owner reference causes problems when the owner is
already borrowed for another purpose.  So the owner reference has been
removed from those calls.  Where the crate user prefers to specify the
owner to get the additional sanity check, a new `owner.cell(val)` call
has been added to support this.

QCell::new() still requires the owner reference (to copy out the owner
ID).  To support creating QCell instances when the QCellOwner is
borrowed, a new QCellOwnerID type is exposed which can be used to
store the ID ready to create new QCells from it, using id.cell(val).