Add `TCellOwner::try_new` and `TCellOwner::wait_for_new`
This commit adds two new functions:
- `TCellOwner::try_new`: checks if a `TCellOwner` already exists,
  and returns `None` if it does.
- `TCellOwner::wait_for_new`: if a `TCellOwner` already exists, it
  blocks the thread until that existing owner is dropped.

Internally, this was implemented using a single static `Condvar`.
If a `wait_for_new` call finds that a `TCellOwner` with its marker
type `Q` already exists, it waits for this `Condvar`. Any time a
`TCellOwner` is dropped, all threads waiting for this `Condvar` are
woken up and check if the `TCellOwner` that was dropped was their
marker type `Q`.

Performance-wise, this means that dropping a `TCellOwner` is now a
tiny bit slower (the `Condvar::notify_all` call). This should be
completely insignificant for any realistic use case. The performance
of every other method (including `TCellOwner::new`) is unaffected.
`TCellOwner::wait_for_new` is just as fast as `TCellOwner::new` if
there is no need to wait.